requirements.txt
livekit-agents>=1.2.0
livekit-plugins-openai
livekit-plugins-silero
python-dotenv
